+++ src/Sav/Record/ValueLabel.php
@@ -151,14 +151,8 @@ class ValueLabel extends Record
         $buffer->writeInt(self::TYPE);
         $buffer->writeInt(\count($this->labels));
         foreach ($this->labels as $item) {
-            $labelLength      = min(mb_strlen($item['label']), self::LABEL_MAX_LENGTH);
-            $label            = mb_substr($item['label'], 0, $labelLength);
-            $labelLengthBytes = mb_strlen($label, '8bit');
-            while ($labelLengthBytes > 255) {
-                // Strip one char, can be multiple bytes
-                $label            = mb_substr($label, 0, -1);
-                $labelLengthBytes = mb_strlen($label, '8bit');
-            }
+            $label = $buffer->encodeString($item['label'], self::LABEL_MAX_LENGTH);
+            $labelLengthBytes = \strlen($label);
 
             if ($convertToDouble) {
                 $item['value'] = Utils::stringToDouble($item['value']);
@@ -166,7 +160,11 @@ class ValueLabel extends Record
 
             $buffer->writeDouble($item['value']);
             $buffer->write(\chr($labelLengthBytes));
-            $buffer->writeString($label, Utils::roundUp($labelLengthBytes + 1, 8) - 1);
+            $buffer->writeString(
+                $label,
+                Utils::roundUp($labelLengthBytes + 1, 8) - 1,
+                $buffer->charset ?? mb_internal_encoding(),
+            );
         }
 
         // Value label variable record.