+++ src/Sav/Record/Variable.php
@@ -156,18 +156,15 @@ class Variable extends Record
         $buffer->writeString($this->name, 8);
 
         if ($hasLabel) {
-            // Maxlength is 255 bytes, since we write utf8 a char can be multiple bytes
-            $labelLength      = min(mb_strlen($this->label), 255);
-            $label            = mb_substr($this->label, 0, $labelLength);
-            $labelLengthBytes = mb_strlen($label, '8bit');
-            while ($labelLengthBytes > 255) {
-                // Strip one char, can be multiple bytes
-                $label            = mb_substr($label, 0, -1);
-                $labelLengthBytes = mb_strlen($label, '8bit');
-            }
+            $label = $buffer->encodeString($this->label, 255);
+            $labelLengthBytes = \strlen($label);
 
             $buffer->writeInt($labelLengthBytes);
-            $buffer->writeString($label, Utils::roundUp($labelLengthBytes, 4));
+            $buffer->writeString(
+                $label,
+                Utils::roundUp($labelLengthBytes, 4),
+                $buffer->charset ?? mb_internal_encoding(),
+            );
         }
 
         // TODO: test