+++ tests/VeryLongStringWriterTest.php
@@ -7,6 +7,8 @@ namespace SPSS\Tests;
 use PHPUnit\Framework\Attributes\DataProvider;
 use SPSS\Sav\Reader;
 use SPSS\Sav\Record\Header;
+use SPSS\Sav\Record\Info\LongVariableNames;
+use SPSS\Sav\Record\Info\MachineInteger;
 use SPSS\Sav\Record\Info\VariableDisplayParam;
 use SPSS\Sav\Record\Info\VeryLongString;
 use SPSS\Sav\Variable;
@@ -20,6 +22,7 @@ final class VeryLongStringWriterTest extends TestCase
         int $compression,
     ): void {
         $longValue = str_repeat("\xC3\x95", 320);
+        $longName = "\xC3\xB5\xC3\xA4\xC3\xB6\xC3\xBC\xC3\xA9_long_text";
         self::assertSame(640, strlen($longValue));
 
         $writer = new Writer([
@@ -30,7 +33,7 @@ final class VeryLongStringWriterTest extends TestCase
             'info' => ['characterEncoding' => 'UTF-8'],
             'variables' => [
                 [
-                    'name' => 'long_text',
+                    'name' => $longName,
                     'format' => Variable::FORMAT_TYPE_A,
                     'width' => 700,
                     'measure' => Variable::MEASURE_ORDINAL,
@@ -50,12 +53,17 @@ final class VeryLongStringWriterTest extends TestCase
             ],
         ]);
 
+        $physicalVariable = $writer->variables[0];
+        self::assertSame("\xC3\x95\xC3\x84", $physicalVariable->name);
+        self::assertSame("\xC3\x95\xC3\x84_A", $physicalVariable->getSegmentName(0));
+        self::assertSame("\xC3\x95\xC3\x84_B", $physicalVariable->getSegmentName(1));
+
         $buffer = $writer->getBuffer();
         $buffer->rewind();
 
         $reader = Reader::fromString($buffer->getStream())->read();
 
-        self::assertSame(['LONG_' => 700], $reader->info[VeryLongString::SUBTYPE]->toArray());
+        self::assertSame(["\xC3\x95\xC3\x84" => 700], $reader->info[VeryLongString::SUBTYPE]->toArray());
         self::assertSame([
             [Variable::MEASURE_ORDINAL, 37, Variable::ALIGN_CENTER],
             [Variable::MEASURE_ORDINAL, 37, Variable::ALIGN_CENTER],
@@ -69,7 +77,7 @@ final class VeryLongStringWriterTest extends TestCase
 
         $buffer->rewind();
         $dataset = Reader::fromString($buffer->getStream())->readDataset();
-        $longVariable = $dataset->variable('long_text');
+        $longVariable = $dataset->variable($longName);
         $shortVariable = $dataset->variable('short_text');
 
         self::assertNotNull($longVariable);
@@ -81,6 +89,51 @@ final class VeryLongStringWriterTest extends TestCase
             [$longValue, 'short'],
             ['', 'text'],
         ], $dataset->rows());
+    }
+
+    #[DataProvider('compressionProvider')]
+    public function testSingleByteEncodingInfoLengthsAndCharacterCodeRoundTrip(
+        string $recordType,
+        int $compression,
+    ): void {
+        $longName = "p\xC3\xB5ld_long_text";
+        $longValue = str_repeat("\xC3\xA4", 300);
+        $writer = new Writer([
+            'header' => [
+                'recType' => $recordType,
+                'compression' => $compression,
+            ],
+            'info' => ['characterEncoding' => 'ISO-8859-1'],
+            'variables' => [[
+                'name' => $longName,
+                'format' => Variable::FORMAT_TYPE_A,
+                'width' => 300,
+                'data' => [$longValue],
+            ]],
+        ]);
+
+        $physicalName = "P\xC3\x95LD_";
+        $physicalVariable = $writer->variables[0];
+        self::assertSame($physicalName, $physicalVariable->name);
+        self::assertSame($physicalName . '_A', $physicalVariable->getSegmentName(0));
+
+        $buffer = $writer->getBuffer();
+        $buffer->rewind();
+
+        $reader = Reader::fromString($buffer->getStream())->read();
+
+        $machineInteger = $reader->info[MachineInteger::SUBTYPE];
+        self::assertInstanceOf(MachineInteger::class, $machineInteger);
+        self::assertSame(28591, $machineInteger->characterCode);
+        self::assertSame(
+            [$physicalName => $longName],
+            $reader->info[LongVariableNames::SUBTYPE]->toArray(),
+        );
+        self::assertSame(
+            [$physicalName => 300],
+            $reader->info[VeryLongString::SUBTYPE]->toArray(),
+        );
+        self::assertSame([[$longValue]], $reader->data);
     }
 
     /** @return iterable<string, array{string, int}> */